Fix blank scene after link navigation (#4964)
Sometimes when navigating using the link feature, the scene would fail
to load, leaving the user in an empty space. The scene was trying to start
rendering upon entering vr, before it was fully loaded. When the render
function was called, it would sometimes pass an undefined camera to the
renderer, causing an exception.

To fix this, I made it check if the scene is loaded before it starts
rendering.